Modularize + add unit tests + verbose flag + glob input
