+++ cli/Cargo.toml
@@ -40,6 +40,7 @@ pem = "3.0.3"
 openssl = { version = "0.10.61", features = ["vendored"] }
 reqwest = { version = "0.12.4", features = ["blocking"] }
 url = "2.5.0"
+glob = "0.3.1"
 
 [dev-dependencies]
 assert_cmd = "2.0.14"
