+++ cli/src/commands/mod.rs
@@ -1,0 +1,5 @@
+mod sign;
+mod view;
+
+pub use sign::sign;
+pub use view::view;
